restructure fixture folders
