+++ tests/fixtures/deploy-fixtures/index.ts
@@ -1,0 +1,12 @@
+export * from './bucketMinimum';
+export * from './bucketWithWebsite';
+export * from './esMinimum';
+export * from './minimum';
+export * from './oneFc';
+export * from './oneFcLargeCode';
+export * from './oneFcOneGateway';
+export * from './oneFcWithContainer';
+export * from './oneFcWithGpu';
+export * from './oneFcWithNas';
+export * from './oneFcWithStage';
+export * from './referredService';
